requirements.txt
Pillow==5.4.1
gevent==1.4.0
python-telegram-bot==11.1.0
